+++ R/data_tbl_functions.R
@@ -74,14 +74,14 @@ los_ra_index_summary_tbl <- function(.data,
                                      .elos_col,
                                      .readmit_rate,
                                      .readmit_bench) {
-    # Tidyeval
+    # * Tidyeval ----
     max_los_var_expr       <- .max_los
     alos_col_var_expr      <- rlang::enquo(.alos_col)
     elos_col_var_expr      <- rlang::enquo(.elos_col)
     readmit_rate_var_expr  <- rlang::enquo(.readmit_rate)
     readmit_bench_var_expr <- rlang::enquo(.readmit_bench)
 
-    # Checks
+    # * Checks ----
     if (!is.data.frame(.data)) {
         stop(call. = FALSE,
              "(data) is not a data-frame/tibble. Please provide.")
@@ -111,7 +111,7 @@ los_ra_index_summary_tbl <- function(.data,
         stop(call. = FALSE, "(.data) has more than 4 columns. Please fix.")
     }
 
-    # Summarize and Manipulate
+    # * Summarize and Manipulate ----
     df_tbl <- tibble::as_tibble(.data) %>%
         dplyr::mutate(alos = {
             {
@@ -159,6 +159,7 @@ los_ra_index_summary_tbl <- function(.data,
         dplyr::mutate(los_ra_var = base::abs(1 - los_index) + base::abs(1 - rar_index)) %>%
         dplyr::select(los_group, los_index, rar_index, los_ra_var)
 
+    # * Return ----
     return(df_summary_tbl)
 
 }
@@ -384,155 +385,3 @@ top_n_tbl <- function(
 
 }
 
-#' Time Series - Census and LOS by Day
-#'
-#' @author Steven P. Sanderson II, MPH
-#'
-#' @description
-#' Sometimes it is important to know what the census was on any given day, or what
-#' the average length of stay is on given day, including for those patients that
-#' are not yet discharged. This can be easily achieved. This will return one
-#' record for every account so the data will still need to be summarized.
-#'
-#' __This function can take a little bit of time to run while the join comparison runs.__
-#'
-#' @details
-#' - Requires a dataset that has at least a start date column and an end date
-#' column
-#' - Takes a single boolean parameter
-#'
-#' @param .data The data you want to pass to the function
-#' @param .keep_nulls_only A boolean that will keep only those records that have
-#' a NULL end date, meaning the patient is still admitted. The default is FALSE which
-#' brings back all records.
-#' @param .start_date_col The column containing the start date for the record
-#' @param .end_date_col The column containing the end date for the record.
-#' @param .by_time How you want the data presented, defaults to day and should remain
-#' that way unless you need more granular data.
-#'
-#' @examples
-#' library(healthyR)
-#' library(healthyR.data)
-#' library(dplyr)
-#'
-#' df <- healthyR_data
-#'
-#' df_tbl <- df %>%
-#'   filter(ip_op_flag == "I") %>%
-#'   select(visit_start_date_time, visit_end_date_time) %>%
-#'   timetk::filter_by_time(.date_var = visit_start_date_time, .start_date = "2020")
-#'
-#' ts_census_los_daily_tbl(
-#'    .data              = df_tbl
-#'    , .keep_nulls_only = FALSE
-#'    , .start_date_col  = visit_start_date_time
-#'    , .end_date_col    = visit_end_date_time
-#' )
-#'
-#' @return
-#' A tibble object
-#'
-#' @export
-#'
-
-ts_census_los_daily_tbl <- function(.data, .keep_nulls_only = FALSE,
-                                    .start_date_col, .end_date_col,
-                                    .by_time = "day"){
-
-    # * Tidyeval Setup ----
-    start_date_var_expr <- rlang::enquo(.start_date_col)
-    end_date_var_expr   <- rlang::enquo(.end_date_col)
-    by_var_expr         <- .by_time
-
-    # * Checks ----
-    if(!is.data.frame(.data)){
-        stop(call. = FALSE,"(.data) is not a data.frame/tibble. Please supply.")
-    }
-
-    if(rlang::quo_is_missing(start_date_var_expr)){
-        stop(call. = FALSE,"(.start_date_col) is missing. Please supply.")
-    }
-
-    if(rlang::quo_is_missing(end_date_var_expr)){
-        stop(call. = FALSE,"(.end_date_col) is missing. Please supply.")
-    }
-
-    keep_nulls_only_bool <- .keep_nulls_only
-
-    # * Data ----
-    data_tbl <- tibble::as_tibble(.data)
-
-    # * Manipulate ----
-    # Get start date and end date
-    all_dates_tbl <- data_tbl %>%
-        dplyr::select(
-            {{ start_date_var_expr }}
-            , {{ end_date_var_expr }}
-        ) %>%
-        purrr::set_names("start_date","end_date") %>%
-        dplyr::mutate(start_date = as.Date(start_date)) %>%
-        dplyr::mutate(end_date   = as.Date(end_date))
-
-    # Filter out all na.dates
-    all_dates_tbl <- all_dates_tbl %>%
-        dplyr::filter(!is.na(start_date)) %>%
-        dplyr::filter(!is.na(end_date))
-
-    start_date <- min(all_dates_tbl[[1]], all_dates_tbl[[2]])
-    end_date   <- max(all_dates_tbl[[1]], all_dates_tbl[[2]])
-    today      <- Sys.Date()
-
-    ts_day_tbl <- timetk::tk_make_timeseries(
-        start_date = start_date
-        , end_date = end_date
-        , by       = by_var_expr
-    ) %>%
-        tibble::as_tibble() %>%
-        dplyr::rename("date"="value") %>%
-        dplyr::mutate(date = as.Date(date))
-
-    res <- sqldf::sqldf(
-        "
-        SELECT *
-        FROM ts_day_tbl AS A
-        LEFT JOIN all_dates_tbl AS B
-        ON start_date <= A.date
-            AND start_date >= A.date
-        "
-    )
-
-    res <- tibble::as_tibble(res) %>%
-        dplyr::arrange(end_date)
-
-    los_tbl <- res %>%
-        dplyr::mutate(
-            los = dplyr::case_when(
-                !is.na(end_date) ~ difftime(
-                    end_date, start_date, units = by_var_expr
-                ) %>% as.integer()
-                , TRUE ~ difftime(
-                    today, start_date, units = by_var_expr
-                ) %>% as.integer()
-            )
-        ) %>%
-        dplyr::mutate(census = 1) %>%
-        # dplyr::mutate(
-        #     date = dplyr::case_when(
-        #         is.na(end_date) ~ Sys.Date()
-        #         , TRUE ~ end_date
-        #     )
-        # )
-        dplyr::arrange(date)
-
-    # Keep NA columns?
-    if(!keep_nulls_only_bool){
-        data_final_tbl <- los_tbl
-    } else {
-        data_final_tbl <- los_tbl %>%
-            dplyr::filter(is.na(end_date))
-    }
-
-    # * Return ----
-    return(data_final_tbl)
-
-}