+++ NEWS.md
@@ -8,6 +8,7 @@ None
 
 ## Minor Fixes and Improvements
 1. Fix #116 - Drop need for `cli`, `crayon`, and `rstudioapi`
+2. F-x #105 - correct code and make minor updates to `ts_census_los_daily_tbl()`
 
 # healthyR 0.1.7